% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscatter.R
\name{rscatter}
\alias{rscatter}
\title{Scalable scatterplot HTML widget}
\usage{
rscatter(x, y, data = NULL, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{x}{numeric vector of x coordinates, OR variable name for x in \code{data}}

\item{y}{numeric vector of y coordinates, OR variable name for y in \code{data}}

\item{data}{optional data.frame containing data to plot}
}
\description{
Create interactive scalable scatterplot using \code{regl-scatterplot} JavaScript library.
}
