% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscatter.R
\name{rscatter}
\alias{rscatter}
\title{Scalable scatterplot HTML widget}
\usage{
rscatter(
  x,
  y,
  size = NULL,
  color = NULL,
  opacity = NULL,
  colorBy = NULL,
  data = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{x}{numeric vector of x coordinates, OR column name for x in \code{data}}

\item{y}{numeric vector of y coordinates, OR column name for y in \code{data}}

\item{size}{point size}

\item{color}{point color}

\item{opacity}{point opacity}

\item{colorBy}{factor/chr/numeric vector to colorBy, OR column name for colorBy in \code{data}}

\item{data}{optional data.frame containing data to plot}

\item{width}{fixed width of canvas in pixels (default is resizable)}

\item{height}{fixed height of canvas in pixels (default is resizable)}

\item{elementId}{specify id for containing div}
}
\description{
Create interactive scalable scatterplot using \code{regl-scatterplot} JavaScript library.
}
\details{
For categorical color encoding ensure the \code{colorBy} values are factor or character vectors.
}
\examples{
data(quakes)
rscatter(quakes$long, quakes$lat, size = 10, colorBy = quakes$depth)

# Or pass data.frame
phi = seq(from = 0, to = 50, by = 1e-2)
fermat_spiral = data.frame(
 x = c(-sqrt(phi)*cos(phi), sqrt(phi)*cos(phi)),
 y = c(-sqrt(phi)*sin(phi), sqrt(phi)*sin(phi)),
 branch = rep(c("a", "b"), each = length(phi))
 )
rscatter("x", "y", size = 1.5, colorBy = "branch", data = fermat_spiral)

}
